Not to be used. Has no API to interact with REACT
This was a second attempt at creating the portal to register a school and add staff members who would have access to the data.
I did not create api app for it to communicate with the frontend. The assumption was that the required html files and css files would be added.